test: only test initial solution for cbc
